Moved code from the namespace & registry auth_scope's to a common base class
