+++ app/controllers/api/base_controller.rb
@@ -1,17 +1,15 @@
 class Api::BaseController < ActionController::Base
   class ScopeNotHandled < StandardError; end
   class RegistryNotHandled < StandardError; end
-  class WrongPortusOTP < StandardError; end
 
   include Pundit
 
   respond_to :json
 
-  rescue_from Namespace::AuthScope::ResourceIsNotFound, with: :deny_access
   rescue_from Pundit::NotAuthorizedError, with: :deny_access
   rescue_from ScopeNotHandled, with: :deny_access
   rescue_from RegistryNotHandled, with: :deny_access
-  rescue_from WrongPortusOTP, with: :deny_access
+  rescue_from Portus::AuthScope::ResourceNotFound, with: :deny_access
 
   protected
