+++ config/application.rb
@@ -23,5 +23,6 @@ module Portus
     # Do not swallow errors in after_commit/after_rollback callbacks.
     config.active_record.raise_in_transactional_callbacks = true
     config.active_record.observers = :user_observer, :registry_observer
+    config.autoload_paths << Rails.root.join("lib")
   end
 end
